requirements.txt
pandas==1.3.0
tensorflow==2.3.1
keras==2.4.3
numpy==1.18.5
matplotlib==3.3.2
pydot==1.4.1
graphviz==0.14.2
seaborn==0.11.0
requests==2.26.0
python-dotenv==0.18.0
apache-airflow==2.1.2
SQLAlchemy==1.3.24
psycopg2-binary==2.9.1
tqdm==4.61.2